Memory worker bounds checking fix RE-229

+++ src/workers/memory/memory_worker.h
@@ -10,9 +10,9 @@ class Memory_Worker : public Worker{
         Memory_Worker(const BehaviourContainer& container, const std::string& inst_name);
         ~Memory_Worker();
         
-        void Allocate(size_t kilobytes);
-        void Deallocate(size_t kilobytes);
-        size_t GetAllocatedCount() const;
+        void Allocate(int kilobytes);
+        void Deallocate(int kilobytes);
+        int GetAllocatedCount() const;
 
     private:
         Memory_Worker_Impl* impl_ = 0;
